+++ app/Filament/Workspace/Resources/KycVerifications/Pages/ListKycVerifications.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Workspace\Resources\KycVerifications\Pages;
+
+use App\Filament\Workspace\Resources\KycVerifications\KycVerificationResource;
+use Filament\Resources\Pages\ListRecords;
+
+class ListKycVerifications extends ListRecords
+{
+    protected static string $resource = KycVerificationResource::class;
+}