+++ bootstrap/providers.php
@@ -3,5 +3,6 @@
 return [
     App\Providers\AppServiceProvider::class,
     App\Providers\Filament\AdminPanelProvider::class,
+    App\Providers\Filament\WorkspacePanelProvider::class,
     App\Providers\TenancyServiceProvider::class,
 ];